finalizing error fix
