+++ src/app/community/study/write/page.tsx
@@ -1,44 +1,13 @@
-'use client';
-
-import { useState, useEffect } from 'react';
-import OnlineForm from './components/OnlineForm';
-import HybridForm from './components/HybridForm';
-import { MEETING_TYPE } from '@/types/study';
-import { useSearchParams } from 'next/navigation';
-
-type StudyType = 'ONLINE' | 'HYBRID';
+import { Suspense } from 'react';
+import StudyWriteForm from '@/app/community/study/write/components/StudyWriteForm';
 
 export default function StudyWritePage() {
-  const [studyType, setStudyType] = useState<StudyType>('ONLINE');
-  const searchParams = useSearchParams();
-
-  useEffect(() => {
-    const type = searchParams.get('type');
-    if (type === 'HYBRID') {
-      setStudyType('HYBRID');
-    }
-  }, [searchParams]);
-
   return (
     <div className="max-w-4xl mx-auto p-4 space-y-6">
       <h1 className="text-2xl font-bold">스터디 모집글 작성</h1>
-
-      <div className="tabs tabs-boxed">
-        <button
-          className={`tab ${studyType === 'ONLINE' ? 'tab-active' : ''}`}
-          onClick={() => setStudyType('ONLINE')}
-        >
-          {MEETING_TYPE.ONLINE}
-        </button>
-        <button
-          className={`tab ${studyType === 'HYBRID' ? 'tab-active' : ''}`}
-          onClick={() => setStudyType('HYBRID')}
-        >
-          {MEETING_TYPE.HYBRID}
-        </button>
-      </div>
-
-      {studyType === 'ONLINE' ? <OnlineForm /> : <HybridForm />}
+      <Suspense fallback={<div>로딩 중...</div>}>
+        <StudyWriteForm />
+      </Suspense>
     </div>
   );
 }
